+++ src/main/java/com/example/one/controller/PublishController.java
@@ -1,8 +1,8 @@
 package com.example.one.controller;
 
 import com.example.one.Dto.QuestionDTO;
-import com.example.one.Model.Question;
-import com.example.one.Model.User;
+import com.example.one.model.Question;
+import com.example.one.model.User;
 import com.example.one.Service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -24,7 +24,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -45,7 +45,7 @@ public class PublishController {
             @RequestParam("description")String description,
             @RequestParam("tag")String tag,
             HttpServletRequest request,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             Model model){
         model.addAttribute("title",title);
         model.addAttribute("description",description);
